main_api_service/requirements.txt
fastapi>=0.111.0,<0.112.0
uvicorn>=0.30.0,<0.31.0
pydantic>=2.7.2,<2.8.0
pydantic[email]
asyncpg>=0.29.0,<0.30.0
psycopg>=3.1.19,<3.2.0
sqlalchemy>=2.0.30,<2.1.0
alembic>=1.13.1,<1.14.0
redis>=5.0.4,<5.1.0
redis[hiredis]
pytest>=8.2.1,<8.3.0
pytest-asyncio>=0.23.7,<0.24.0
pytest-mock>=3.14.0,<3.15.0
httpx>=0.27.0,<0.28.0
aiokafka>=0.10.0,<0.11.0
kafka-python3>=3.0.0,<3.1.0
PyJWT>=2.8.0,<2.9.0
weasyprint>=62.1,<63.0
img2pdf>=0.5.1,<0.5.2
imageio>=2.34.1,<2.35.0
python-multipart>=0.0.9,<0.0.10
argon2-cffi>=23.1.0,<23.2.0